feat: rename and rearrange package
